Added some validations & settings for post request

+++ watchStewar/watchStewar.Common/Models/Watch.cs
@@ -8,7 +8,7 @@ namespace watchStewar.Common.Models
 
         public DateTime register { get; set; }
 
-        public byte type { get; set; }
+        public int type { get; set; }
 
         public bool isConsolidate { get; set; }
     }
